+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ x.x.x (x-x-x)
 * Removed support for Python 3.6 (end of life since December 2021).
 * Added support for Python 3.10 and Python 3.11.
 * Added the option to keep temporary files for app push and pull.
+* Added logout command.
 
 3.17.2 (2023-03-16)
 -------------------